Continued working on Lesson 8 HW day 5, added UI

+++ src/main/java/lv/javaguru/java1/student_anton_pereloma/lesson_8/homework/day_5/Merchandise.java
@@ -51,7 +51,7 @@ class Merchandise {
         return reviews.get(reviewNumber);
     }
 
-    public void newReview(String author, String reviewText, double rating) {
+    public void newReview(String author, String reviewText, int rating) {
         Review newReview = new Review(author,reviewText,rating);
         reviews.add(newReview);
     }
